# Conflicts:
#	JitHub/JitHub.csproj
#	JitHub/Services/GitHubService.cs
#	JitHub/Services/IGitHubService.cs
#	JitHub/Views/Controls/Widgets/RepoSideBar.xaml

+++ JitHub/Services/AccountService.cs
@@ -15,7 +15,7 @@
             _settings.Save(userIdKey, string.Empty);
         }
 
-        public void SaveUser(int userId)
+        public void SaveUser(long userId)
         {
             _settings.Save(userIdKey, userId);
             _settings.Save(doNotWarnDeleteRepoKey, false);
